+++ src/expand/mod.rs
@@ -413,9 +413,9 @@ where
             i.iter_elements().map(AsChar::as_char).eq(str.chars())
         };
 
-        if eq(&self.0, "int") {
-            Left(r#"((?:-?\d+)|(?:\d+))"#.chars().map(Ok))
-        } else if eq(&self.0, "float") {
+        if eq(&self.input, "int") {
+            Left(Left(r#"((?:-?\d+)|(?:\d+))"#.chars().map(Ok)))
+        } else if eq(&self.input, "float") {
             // Regex in other implementations has lookaheads. As `regex` crate
             // doesn't support them, we use `f32`/`f64` grammar instead:
             // https://doc.rust-lang.org/stable/std/primitive.f64.html#grammar
@@ -423,26 +423,36 @@ where
             // - supports `e` as exponent in addition to `E`
             // - supports trailing comma: `1.`
             // - supports `inf` and `NaN`
-            Left(
+            Left(Left(
                 "([+-]?(?:inf\
                          |NaN\
                          |(?:\\d+|\\d+\\.\\d*|\\d*\\.\\d+)(?:[eE][+-]?\\d+)?\
                        ))"
                 .chars()
                 .map(Ok),
-            )
-        } else if eq(&self.0, "word") {
-            Left(r#"([^\s]+)"#.chars().map(Ok))
-        } else if eq(&self.0, "string") {
-            Left(
-                r#"("(?:[^"\\]*(?:\\.[^"\\]*)*)"|'(?:[^'\\]*(?:\\.[^'\\]*)*)')"#
-                    .chars()
-                    .map(Ok),
-            )
-        } else if eq(&self.0, "") {
-            Left(r#"(.*)"#.chars().map(Ok))
+            ))
+        } else if eq(&self.input, "word") {
+            Left(Left(r#"([^\s]+)"#.chars().map(Ok)))
+        } else if eq(&self.input, "string") {
+            // We add `string_<id>` name to `string` parameters to later parse
+            // out leading and trailing quotes in `cucumber-codegen` crate.
+            // See https://github.com/cucumber-rs/cucumber-expressions/issues/7
+            // for more info.
+            Left(Right(
+                OwnedChars::new(format!(
+                    "(?P<string_{}>\
+                      \"(?:[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"\
+                      |'(?:[^'\\\\]*(?:\\\\.[^'\\\\]*)*)')",
+                    self.id,
+                ))
+                .map(Ok),
+            ))
+        } else if eq(&self.input, "") {
+            Left(Left(r#"(.*)"#.chars().map(Ok)))
         } else {
-            Right(iter::once(Err(UnknownParameterError { not_found: self.0 })))
+            Right(iter::once(Err(UnknownParameterError {
+                not_found: self.input,
+            })))
         }
     }
 }
@@ -451,9 +461,15 @@ where
 //       https://github.com/rust-lang/rust/issues/63063
 /// [`IntoRegexCharIter::Iter`] for a [`Parameter`].
 type ParameterIter<Input> = Either<
-    iter::Map<
-        str::Chars<'static>,
-        fn(char) -> Result<char, UnknownParameterError<Input>>,
+    Either<
+        iter::Map<
+            str::Chars<'static>,
+            fn(char) -> Result<char, UnknownParameterError<Input>>,
+        >,
+        iter::Map<
+            OwnedChars,
+            fn(char) -> Result<char, UnknownParameterError<Input>>,
+        >,
     >,
     iter::Once<Result<char, UnknownParameterError<Input>>>,
 >;
@@ -510,6 +526,33 @@ where
     fn next(&mut self) -> Option<Self::Item> {
         let next = self.iter.next();
         (self.iter.peek().is_some()).then(|| next).flatten()
+    }
+}
+
+/// Like [`str::Chars`], but owns [`String`].
+#[derive(Clone, Debug)]
+pub struct OwnedChars {
+    /// Iterated [`String`].
+    str: String,
+
+    /// Current char number.
+    cur: usize,
+}
+
+impl OwnedChars {
+    /// Creates a new [`OwnedChars`] [`Iterator`].
+    pub fn new(str: String) -> Self {
+        Self { str, cur: 0 }
+    }
+}
+
+impl Iterator for OwnedChars {
+    type Item = char;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let char = self.str.chars().nth(self.cur)?;
+        self.cur += 1;
+        Some(char)
     }
 }
 
@@ -699,7 +742,9 @@ mod spec {
 
         assert_eq!(
             expr.as_str(),
-            r#"^("(?:[^"\\]*(?:\\.[^"\\]*)*)"|'(?:[^'\\]*(?:\\.[^'\\]*)*)')$"#,
+            "^(?P<string_0>\
+               \"(?:[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
+               '(?:[^'\\\\]*(?:\\\\.[^'\\\\]*)*)')$",
         );
         assert!(expr.is_match("\"\""));
         assert!(expr.is_match("''"));
@@ -708,6 +753,29 @@ mod spec {
         assert!(expr.is_match("\"with \\\" escaped\""));
         assert!(expr.is_match("'with \\' escaped'"));
         assert!(!expr.is_match("word"));
+    }
+
+    #[test]
+    fn multiple_string_parameters() {
+        // TODO: Use "{e}" syntax once MSRV bumps above 1.58.
+        let expr = Expression::regex("{string} {string}")
+            .unwrap_or_else(|e| panic!("failed: {}", e));
+
+        assert_eq!(
+            expr.as_str(),
+            "^(?P<string_0>\
+               \"(?:[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
+               '(?:[^'\\\\]*(?:\\\\.[^'\\\\]*)*)') \
+             (?P<string_1>\
+               \"(?:[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
+               '(?:[^'\\\\]*(?:\\\\.[^'\\\\]*)*)')$",
+        );
+        assert!(expr.is_match("\"\" ''"));
+        assert!(expr.is_match("'' \"\""));
+        assert!(expr.is_match("'with \"' \"\""));
+        assert!(expr.is_match("\"with '\" '\"'"));
+        assert!(expr.is_match("\"with \\\" escaped\" 'with \\' escaped'"));
+        assert!(expr.is_match("'with \\' escaped' \"with \\\" escaped\""));
     }
 
     #[test]